# app/Http/Requests/Api/Admin/Platform/CreatePlatformRequest.php
<?php

namespace App\Http\Requests\Api\Admin\Platform;

use Illuminate\Foundation\Http\FormRequest;

/**  
 * @OA\Schema(  
 *     schema="CreatePlatformRequest",  
 *     type="object",  
 *     required={  
 *         "name"  
 *     },  
 *     @OA\Property(property="name", type="string", maxLength=255, example="Sample Name"),  
 * )  
 */
class CreatePlatformRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'name' => 'required|string|max:255',
        ];
    }
}