# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GUG Apis ",
        "contact": {
            "name": "Yosof Bayan",
            "url": "https://wa.me/+963967213544",
            "email": "yosofbayan75@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "local Base URL"
        },
        {
            "url": "http://gug-back.icrcompany.net/api/v1",
            "description": "Develop Base URL"
        }
    ],
    "paths": {
...
        "/admin/platforms": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Platform"
                ],
                "summary": "get all platforms",
                "operationId": "0fc715a8da3e5dfda7d9002f1ddb4a65",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - Platform"
                ],
                "summary": "Create a new Platform",
                "operationId": "443b8605138a87141a3491e18a45ebe1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePlatformRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/admin/platforms/{platform}": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Platform"
                ],
                "summary": "show a Platform",
                "operationId": "771f0bd6530c5e1ecd14f9220b10449d",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - Platform"
                ],
                "summary": "update an existing  Platform",
                "operationId": "5c6e1e78f9edb048882b56def8dc4e1c",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT",
                            "enum": [
                                "PUT"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePlatformRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
...
        "/admin/sections/{section}/projects": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Section - Project"
                ],
                "summary": "get all sections/{section}/projects",
                "operationId": "9b19aef28fc84c344f03759b60465382",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - Section - Project"
                ],
                "summary": "Create a new Section - Project",
                "operationId": "9aa2802ca4f0d9258fa965836fda7ecb",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/admin/sections/{section}/projects/{project}": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Section - Project"
                ],
                "summary": "show an project",
                "operationId": "8930889f0463d6cd21c3666ff12638c3",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - Section - Project"
                ],
                "summary": "update an existing  Section - Project",
                "operationId": "746c99a3a358f74aa56110aff8199f4d",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT",
                            "enum": [
                                "PUT"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
...
        "/platforms": {
            "get": {
                "tags": [
                    "App",
                    "App - Platform"
                ],
                "summary": "get all section platforms",
                "operationId": "3a8eb79c5fea9287437431a690670b81",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
...
        "/sections/{section}/projects": {
            "get": {
                "tags": [
                    "App",
                    "App - Section - Project"
                ],
                "summary": "get all sections/{section}/projects",
                "operationId": "303200789f9c3c9ee3f174a912b61c9d",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/sections/all/projects": {
            "get": {
                "tags": [
                    "App",
                    "App - Section - Project"
                ],
                "summary": "get all sections/{section}/projects",
                "operationId": "bcd247793ba2b9123c67f4271b3fea52",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/admin/sections/all/projects": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Section - Project"
                ],
                "summary": "get all sections/{section}/projects",
                "operationId": "215da73c111662571f29dbaa8cece315",
                "parameters": [
                    {
                        "name": "trash",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/sections/{section}/projects/{project}": {
            "get": {
                "tags": [
                    "App",
                    "App - Section - Project"
                ],
                "summary": "show an project",
                "operationId": "2ab185059e08bdc728a7526a6d83291d",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreatePlatformRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    }
                },
                "type": "object"
            },
            "UpdatePlatformRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    }
                },
                "type": "object"
            },
...
            "CreateProjectRequest": {
                "required": [
                    "title",
                    "description",
                    "logo"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "description": {},
                    "logo": {
                        "type": "string",
                        "format": "binary",
                        "example": "image.jpg"
                    },
                    "images[0]": {
                        "description": "The logo to be uploaded. Accepts image : jpeg, jpg, png",
                        "type": "string",
                        "format": "binary",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "UpdateProjectRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "description": {},
                    "logo": {
                        "type": "string",
                        "format": "binary",
                        "example": "image.jpg"
                    },
                    "images[0]": {
                        "description": "The logo to be uploaded. Accepts image : jpeg, jpg, png",
                        "type": "string",
                        "format": "binary",
                        "nullable": false
                    },
                    "delete_images[0]": {
                        "description": "The logo to be uploaded. Accepts image : jpeg, jpg, png",
                        "type": "integer",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}